+++ tests/qttools/obc/conftest.py
@@ -10,6 +10,25 @@ from quatrex.core.quatrex_config import parse_config as parse_quatrex_config
 from quatrex.electron.solver import ElectronSolver
 from quatrex.examples import get_example_dir
 
+X_II_FORMULAS = ["self-energy", "direct"]
+
+BLOCK_SIZE = [
+    pytest.param(21, id="21x21"),
+    pytest.param(18, id="18x18"),
+]
+
+BLOCK_SECTIONS = [
+    pytest.param(1, id="no-subblocks"),
+    pytest.param(3, id="three-subblocks"),
+]
+
+BATCH_SIZE = [
+    pytest.param(1, id="single-batch"),
+    pytest.param(3, id="three-batches"),
+]
+
+CONTACTS = ["left", "right"]
+
 NEVP_SOLVERS = [
     pytest.param(
         Beyn(r_o=10, r_i=0.99, m_0=32, num_quad_points=15, use_qr=False), id="Beyn"
@@ -21,6 +40,36 @@ NEVP_SOLVERS = [
 ]
 
 ENERGIES = [[-10, -5, 0]]
+
+
+@pytest.fixture(params=X_II_FORMULAS)
+def x_ii_formula(request: pytest.FixtureRequest) -> str:
+    """Returns a NEVP solver."""
+    return request.param
+
+
+@pytest.fixture(params=BLOCK_SIZE)
+def block_size(request: pytest.FixtureRequest) -> int:
+    """Returns the block size."""
+    return request.param
+
+
+@pytest.fixture(params=BLOCK_SECTIONS)
+def block_sections(request: pytest.FixtureRequest) -> int:
+    """Returns the number of block sections."""
+    return request.param
+
+
+@pytest.fixture(params=BATCH_SIZE)
+def batch_size(request: pytest.FixtureRequest) -> int:
+    """Returns the block size."""
+    return request.param
+
+
+@pytest.fixture(params=CONTACTS, autouse=True)
+def contact(request: pytest.FixtureRequest) -> str:
+    """Returns a contact."""
+    return request.param
 
 
 @pytest.fixture(params=ENERGIES, autouse=True, scope="session")
